# specs/001-b2c-booking-backend/contracts/openapi.yaml
openapi: 3.0.1
info:
  title: B2C 預約系統 API
  description: 用於管理服務、可用時段和預約的 API。
  version: v1.0
paths:
...
  /api/customers/providers/{providerId}/availability:
    get:
      tags:
        - CustomerAvailability
      summary: (Customer) 查詢 Provider 在特定日期範圍內的可用時段
      parameters:
        - name: providerId
          in: path
          required: true
          schema:
            type: integer
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: 成功返回可用時段列表。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailabilitySlotDto'
...
components:
  schemas:
    AvailabilitySlotDto:
      type: object
      properties:
        id:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
...